+++ geofileops/_compat.py
@@ -1,5 +1,7 @@
 """Module with compatibility and dependency availability checks."""
 
+import sys
+
 import geopandas as gpd
 import pandas as pd
 import pyogrio
@@ -17,7 +19,7 @@ gdal.UseExceptions()
 GDAL_BASE_VERSION = version.parse(gdal.__version__.split("-")[0]).base_version
 GDAL_GTE_39 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.9")
 GDAL_GTE_3101 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.10.1")
-GDAL_ST_311 = version.parse(GDAL_BASE_VERSION) < version.parse("3.11")
+GDAL_GTE_311 = version.parse(GDAL_BASE_VERSION) >= version.parse("3.11")
 
 GEOPANDAS_GTE_10 = version.parse(gpd.__version__) >= version.parse("1.0")
 GEOPANDAS_110 = version.parse(gpd.__version__) == version.parse("1.1.0")
@@ -25,6 +27,7 @@ GEOS_GTE_313 = version.parse(shapely.geos_version_string) >= version.parse("3.13
 PANDAS_GTE_20 = version.parse(pd.__version__) >= version.parse("2.0")
 PANDAS_GTE_22 = version.parse(pd.__version__) >= version.parse("2.2")
 PYOGRIO_GTE_010 = version.parse(pyogrio.__version__) >= version.parse("0.10")
+PYTHON_313 = sys.version_info >= (3, 13) and sys.version_info < (3, 14)
 SHAPELY_GTE_20 = version.parse(shapely.__version__) >= version.parse("2")
 
 sqlite3_spatialite_version_info = _sqlite_util.spatialite_version_info()