+++ tests/general_file_layer_operations/test_geofile.py
@@ -18,11 +18,15 @@ from pygeoops import GeometryType
 
 import geofileops as gfo
 from geofileops import _compat, fileops
-from geofileops.util import _geofileinfo, _geopath_util, _geoseries_util
+from geofileops._compat import GDAL_GTE_311
+from geofileops.util import _geofileinfo, _geoseries_util
+from geofileops.util._geopath_util import GeoPath
 from tests import test_helper
 from tests.test_helper import (
     SUFFIXES_FILEOPS,
     SUFFIXES_FILEOPS_EXT,
+    SUFFIXES_GEOOPS,
+    SUFFIXES_GEOOPS_EXT,
     assert_geodataframe_equal,
 )
 
@@ -1353,7 +1357,7 @@ def test_listlayers_one_layer(suffix, only_spatial_layers, expected):
     src = test_helper.get_testfile("polygon-parcel", suffix=suffix)
     layers = gfo.listlayers(src, only_spatial_layers=only_spatial_layers)
 
-    expected = [exp.format(src_stem=_geopath_util.stem(src)) for exp in expected]
+    expected = [exp.format(src_stem=GeoPath(src).stem) for exp in expected]
     assert layers == expected
 
 
@@ -1438,7 +1442,7 @@ def test_update_column_error(tmp_path):
 
     # Trying to update column that doesn't exist should raise ValueError
     assert "not_existing column" not in layerinfo.columns
-    with pytest.raises(ValueError, match="Column .* doesn't exist in"):
+    with pytest.raises(ValueError, match=r"Column .* doesn't exist in"):
         gfo.update_column(
             test_path, name="not_existing column", expression="ST_area(geom)"
         )
@@ -1874,9 +1878,18 @@ def test_to_file(tmp_path, suffix, dimensions, engine_setter):
     """
     # Remark: geopandas doesn't seem seem to read the Z dimension, so writing can't be
     # tested?
-    # Read the file from GPKG, so the dataframe to be written is constent.
-    src = test_helper.get_testfile("polygon-parcel", dimensions=dimensions)
-    read_gdf = gfo.read_file(src)
+    # Prepare test file
+    src = test_helper.get_testfile(
+        "polygon-parcel", suffix=suffix, dimensions=dimensions
+    )
+    uidn = str(2318781) if suffix == ".csv" else 2318781
+    encoding = "utf-8" if suffix == ".csv" else None
+
+    # Read test file and write to tmppath
+    read_gdf = gfo.read_file(src, encoding=encoding)
+
+    # Validate if string (encoding) is correct for data read.
+    assert read_gdf.loc[read_gdf["UIDN"] == uidn]["LBLHFDTLT"].item() == "Silomaïs"
 
     if suffix in (".gpkg.zip", ".shp.zip"):
         pytest.xfail("writing a dataframe to gpkg.zip or .shp.zip has issue")
@@ -1894,7 +1907,7 @@ def test_to_file(tmp_path, suffix, dimensions, engine_setter):
             "well"
         )
 
-    output_path = tmp_path / f"{_geopath_util.stem(src)}-output{suffix}"
+    output_path = tmp_path / f"{GeoPath(src).stem}-output{suffix}"
     gfo.to_file(read_gdf, str(output_path))
 
     kwargs = {}
@@ -1983,7 +1996,7 @@ def test_to_file_2_roundtrip(tmp_path, suffix, dimensions, engine_setter):
     if suffix == ".shp" and engine_setter == "fiona":
         pytest.xfail("fiona writes datetimes to shapefile as Date, loosing data")
 
-    output_path = tmp_path / f"{_geopath_util.stem(src)}-output{suffix}"
+    output_path = tmp_path / f"{GeoPath(src).stem}-output{suffix}"
     gfo.to_file(read_gdf, str(output_path))
     written_gdf = gfo.read_file(output_path)
 
@@ -2040,7 +2053,7 @@ def test_to_file_append(tmp_path, suffix, engine_setter):
     if suffix == ".shp" and engine_setter == "fiona":
         pytest.xfail("fiona does not support writing datetimes to shapefile")
 
-    output_path = tmp_path / f"{_geopath_util.stem(src)}-output{suffix}"
+    output_path = tmp_path / f"{GeoPath(src).stem}-output{suffix}"
     gfo.to_file(test_gdf, path=output_path)
     gfo.to_file(test_gdf, path=output_path, append=True)
 
@@ -2565,6 +2578,24 @@ def test_remove(tmp_path, suffix):
     assert not src.exists()
 
 
+@pytest.mark.parametrize("suffix", SUFFIXES_GEOOPS)
+@pytest.mark.skipif(not GDAL_GTE_311, reason="sozip requires gdal>=3.11")
+def test_zip_geofile(tmp_path, suffix):
+    input_path = test_helper.get_testfile("polygon-parcel", suffix=suffix)
+
+    sozip_path = tmp_path / "zipped.zip"
+    fileops.zip_geofile(input_path, sozip_path)
+
+    # Check result
+    assert sozip_path.exists()
+    assert sozip_path.stat().st_size > 0
+
+    unzipped_dir = tmp_path / "unzipped"
+    fileops.unzip_geofile(sozip_path, unzipped_dir)
+    for path in input_path if isinstance(input_path, list) else [input_path]:
+        assert (unzipped_dir / path.name).exists()
+
+
 def test_launder_columns():
     columns = [f"TOO_LONG_COLUMNNAME{index}" for index in range(21)]
     laundered = fileops._launder_column_names(columns)
@@ -2592,35 +2623,70 @@ def test_launder_columns():
         laundered = fileops._launder_column_names(columns)
 
 
-def test_zip_unzip(tmp_path):
+@pytest.mark.parametrize("suffix, exp_nb_files", [(".gpkg", 1), (".shp", 4)])
+def test_unzip_geofile(tmp_path, suffix, exp_nb_files):
     # Prepare test data
-    src = test_helper.get_testfile("polygon-parcel")
-    zip_path = tmp_path / "zipped.zip"
-    fileops._zip(src, zip_path)
-
-    # Unzip and check result
-    dst_dir = tmp_path / "unzipped"
-    fileops._unzip(zip_path, dst_dir)
-    assert len(list(dst_dir.iterdir())) == 1
-    assert (dst_dir / src.name).exists()
-
-
-def test_zip_unzip_dir(tmp_path):
-    # Prepare test data
-    src = test_helper.get_testfile("polygon-parcel")
+    input_path = test_helper.get_testfile("polygon-parcel", suffix=suffix)
     zip_dir = tmp_path / "dir_to_zip"
     zip_dir.mkdir()
-    file1 = zip_dir / f"{src.stem}_1{src.suffix}"
-    file2 = zip_dir / f"{src.stem}_2{src.suffix}"
-    gfo.copy(src, file1)
-    gfo.copy(src, file2)
+    gfo.copy(input_path, zip_dir)
     zip_path = tmp_path / "zipped.zip"
     fileops._zip(zip_dir, zip_path)
 
     # Unzip and check result
-    dst_dir = tmp_path / "unzipped"
-    fileops._unzip(zip_path, dst_dir)
-    assert dst_dir.exists()
-    assert len(list(dst_dir.iterdir())) == 2
-    assert (dst_dir / file1.name).exists()
-    assert (dst_dir / file2.name).exists()
+    output_dir = tmp_path / "unzipped"
+    output_geofile_path = fileops.unzip_geofile(zip_path, output_dir)
+    assert output_dir.exists()
+    assert output_geofile_path.exists()
+    assert output_geofile_path == output_dir / input_path.name
+    assert len(list(output_dir.iterdir())) == exp_nb_files
+    assert (output_dir / input_path.name).exists()
+
+
+@pytest.mark.parametrize("suffix", SUFFIXES_GEOOPS_EXT)
+def test_unzip_geofile_error_multi_files(tmp_path, suffix):
+    # Prepare test data
+    input_path = test_helper.get_testfile("polygon-parcel", suffix=suffix)
+    zip_dir = tmp_path / "dir_to_zip"
+    zip_dir.mkdir()
+    file = zip_dir / GeoPath(input_path).with_stem_suffix("_1").name
+    gfo.copy(input_path, file)
+    file = zip_dir / GeoPath(input_path).with_stem_suffix("_2").name
+    gfo.copy(input_path, file)
+    zip_path = tmp_path / "zipped.zip"
+    fileops._zip(zip_dir, zip_path)
+
+    # Unzip and check result
+    output_dir = tmp_path / "unzipped"
+    with pytest.raises(ValueError, match="Multiple geofiles found in zip"):
+        _ = fileops.unzip_geofile(zip_path, output_dir)
+
+
+def test_unzip_geofile_error_no_files(tmp_path):
+    # Prepare test data
+    zip_dir = tmp_path / "dir_to_zip"
+    zip_dir.mkdir()
+    zip_path = tmp_path / "zipped.zip"
+    fileops._zip(zip_dir, zip_path)
+
+    # Unzip and check result
+    output_dir = tmp_path / "unzipped"
+    with pytest.raises(ValueError, match="No files found in zip"):
+        _ = fileops.unzip_geofile(zip_path, output_dir)
+
+
+def test_unzip_geofile_error_no_geofiles(tmp_path):
+    # Prepare test data
+    zip_dir = tmp_path / "dir_to_zip"
+    zip_dir.mkdir()
+    file1 = zip_dir / "no_geo_file1.txt"
+    file1.touch()
+    file2 = zip_dir / "no_geo_file2.txt"
+    file2.touch()
+    zip_path = tmp_path / "zipped.zip"
+    fileops._zip(zip_dir, zip_path)
+
+    # Unzip and check result
+    output_dir = tmp_path / "unzipped"
+    with pytest.raises(ValueError, match="No geofile found in zip"):
+        _ = fileops.unzip_geofile(zip_path, output_dir)